# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/api/images/process": {
      "post": {
        "description": "Upload and process an image to remove background. Requires active subscription and credits.",
        "operationId": "ImagesController_processImage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Image file to process",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file (jpg, jpeg, png, gif, bmp, tiff, webp) - Max size: 20MB"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image processing started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Image processing started successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "uuid-string"
                        },
                        "ownerId": {
                          "type": "string",
                          "nullable": true,
                          "example": "user-id-123"
                        },
                        "processId": {
                          "type": "string",
                          "example": "process-uuid"
                        },
                        "originalFileName": {
                          "type": "string",
                          "example": "image.jpg"
                        },
                        "originalImageUrlLQ": {
                          "type": "string",
                          "nullable": true,
                          "example": "https://example.com/image-lq.jpg"
                        },
                        "originalImageUrlHQ": {
                          "type": "string",
                          "nullable": true,
                          "example": "https://example.com/image-hq.jpg"
                        },
                        "bgRemovedFileName": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        },
                        "bgRemovedImageUrlLQ": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        },
                        "bgRemovedImageUrlHQ": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "queue",
                            "processing",
                            "ready"
                          ],
                          "example": "queue"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2023-09-17T10:00:00Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2023-09-17T10:00:00Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid file format or size",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Only image files are allowed!"
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment required - Insufficient credits or subscription required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Insufficient credits or active subscription required"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too many requests - Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Rate limit exceeded"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Process uploaded image",
        "tags": [
          "Images"
        ]
      }
    },
    "/v1/api/images/user/{id}": {
      "get": {
        "description": "Retrieve all images for a specific user with pagination and filtering options",
        "operationId": "ImagesController_findAllImagesByUserId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID to fetch images for",
            "schema": {
              "example": "user-id-123",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items to return per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for pagination (next page)",
            "schema": {
              "example": "cursor-string",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter images by filename",
            "schema": {
              "example": "image-name",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter images by processing status",
            "schema": {
              "enum": [
                "queue",
                "processing",
                "ready"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User images retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Images retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "uuid-string"
                          },
                          "ownerId": {
                            "type": "string",
                            "nullable": true,
                            "example": "user-id-123"
                          },
                          "processId": {
                            "type": "string",
                            "example": "process-uuid"
                          },
                          "originalFileName": {
                            "type": "string",
                            "example": "image.jpg"
                          },
                          "originalImageUrlLQ": {
                            "type": "string",
                            "nullable": true,
                            "example": "https://example.com/image-lq.jpg"
                          },
                          "originalImageUrlHQ": {
                            "type": "string",
                            "nullable": true,
                            "example": "https://example.com/image-hq.jpg"
                          },
                          "bgRemovedFileName": {
                            "type": "string",
                            "nullable": true,
                            "example": "processed-image.png"
                          },
                          "bgRemovedImageUrlLQ": {
                            "type": "string",
                            "nullable": true,
                            "example": "https://example.com/processed-lq.png"
                          },
                          "bgRemovedImageUrlHQ": {
                            "type": "string",
                            "nullable": true,
                            "example": "https://example.com/processed-hq.png"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "queue",
                              "processing",
                              "ready"
                            ],
                            "example": "ready"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2023-09-17T10:00:00Z"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2023-09-17T10:00:00Z"
                          }
                        }
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "limit": {
                          "type": "number",
                          "example": 10
                        },
                        "count": {
                          "type": "number",
                          "example": 25
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": true
                        },
                        "nextCursor": {
                          "type": "string",
                          "nullable": true,
                          "example": "next-cursor-string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Authentication required"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user images",
        "tags": [
          "Images"
        ]
      }
    },
...
  },
  "info": {
    "title": "Erazor Server",
    "description": "The Erazor Server API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}